[1419] Movedparser into own module
